requirements.txt
pandas
requests
openpyxl
matplotlib
pytest
google-cloud-bigquery
pandas-gbq
pyarrow
lxml
html5lib
beautifulsoup4
openpyxl